requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
bcrypt==4.1.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.109.0
greenlet==3.2.4
httptools==0.7.1
httpx==0.24.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.35.1
supabase==2.0.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets>=11,<13